moved App to components folder
